refactor: extract useSessionSync composable from App.vue
refreshSessionStates + markSessionRead + sessions channel pubsub.
